// src/app/models/cicloFilho.ts
export class CicloFilho {
    
    constructor(){ }

    id: number;
    cicloId: number;
    dataInicio: Date;
    dataTermino: Date;
    valorRacao: number;
    valorMedicamentos: number;
    valorGastosExtra: number;
    quantidadeDias: number;
    quantidadeAnimais: number;
    quantidadeRacao: number;
    valor: number;
    tipo: string;
}